feat(meteo): split the multi-source weather integration back into separate Météo France and OpenWeather integrations
