+++ front/src/config/integrations/weathers.json
@@ -6,8 +6,8 @@
     "cloud": true
   },
   {
-    "key": "meteo",
-    "link": "meteo",
+    "key": "meteoFrance",
+    "link": "meteofrance",
     "img": "/assets/integrations/cover/meteo-france.jpg",
     "whiteBackground": true,
     "cloud": true
