+++ server/lib/weatherWidget/openweather.formatter.js
