+++ server/services/meteofrance/controllers/meteofrance.controller.js
@@ -3,17 +3,16 @@ const asyncMiddleware = require('../../../api/middlewares/asyncMiddleware');
 const { parseAlerts, parseVigilanceText } = require('../lib/vigilance.parser');
 
 /**
- * @description Meteo service controllers.
+ * @description MeteoFrance service controllers.
  * @param {any} gladys - Gladys instance.
  * @param {Function} getVigilance - Vigilance fetch function.
- * @param {Function} getForecast - Forecast fetch function (source-aware, Météo France format).
+ * @param {Function} getForecast - Forecast fetch function.
  * @param {Function} getVigilanceMap - Vigilance map fetch function (optional API key).
- * @param {Function} getSource - Returns the configured forecast source.
  * @returns {object} Controllers routes map.
  */
-module.exports = function MeteoController(gladys, getVigilance, getForecast, getVigilanceMap, getSource) {
+module.exports = function MeteoFranceController(gladys, getVigilance, getForecast, getVigilanceMap) {
   /**
-   * @api {get} /api/v1/service/meteo/vigilance Get vigilance alerts
+   * @api {get} /api/v1/service/meteofrance/vigilance Get vigilance alerts
    * @param {any} req - Express request.
    * @param {any} res - Express response.
    */
@@ -28,7 +27,7 @@ module.exports = function MeteoController(gladys, getVigilance, getForecast, get
   }
 
   /**
-   * @api {get} /api/v1/service/meteo/vigilance/map Get national vigilance map image
+   * @api {get} /api/v1/service/meteofrance/vigilance/map Get national vigilance map image
    * @param {any} req - Express request.
    * @param {any} res - Express response.
    */
@@ -42,14 +41,13 @@ module.exports = function MeteoController(gladys, getVigilance, getForecast, get
   }
 
   /**
-   * @api {get} /api/v1/house/:house_selector/meteo/weather Get weather for a house (configured source)
+   * @api {get} /api/v1/house/:house_selector/meteofrance/weather Get Météo France weather for a house
    * @param {any} req - Express request.
    * @param {any} res - Express response.
    */
   async function getHouseWeatherController(req, res) {
     const houseSelector = req.params.house_selector;
     const vigilanceRequested = `${req.query.vigilance}` === 'true';
-    const source = getSource();
 
     const house = await gladys.house.getBySelector(houseSelector);
     if (house.latitude == null || house.longitude == null) {
@@ -61,22 +59,14 @@ module.exports = function MeteoController(gladys, getVigilance, getForecast, get
     try {
       forecastData = await getForecast(house.latitude, house.longitude);
     } catch (e) {
-      // @ts-ignore: custom field set by the service when the OpenWeather key is missing
-      if (e && e.code === 'OPENWEATHER_API_KEY_MISSING') {
-        res.status(400).json({ message: 'OPENWEATHER_API_KEY_MISSING' });
-        return;
-      }
       const detail = e instanceof Error ? e.message : `${e}`;
-      logger.warn(`[Meteo] getForecast (${source}) failed for (${house.latitude},${house.longitude}): ${detail}`);
+      logger.warn(`[MeteoFrance] getForecast failed for (${house.latitude},${house.longitude}): ${detail}`);
       res.status(502).json({ message: 'FORECAST_API_ERROR', detail });
       return;
     }
 
-    // Vigilance is a Météo France feature: it is only fetched with the Météo France
-    // source (its forecast response already contains the department). With the
-    // OpenWeather source, nothing is fetched from Météo France at all.
-    /** @type {string|null} */
-    const dept = (source === 'meteofrance' && forecastData.position && forecastData.position.dept) || null;
+    // The department is derived from the house coordinates through the forecast response
+    const dept = (forecastData.position && forecastData.position.dept) || null;
     /** @type {Array<object>} */
     let alerts = [];
     let text = '';
@@ -87,31 +77,29 @@ module.exports = function MeteoController(gladys, getVigilance, getForecast, get
         if (alerts.length > 0) {
           text = parseVigilanceText(warningData);
         }
-        logger.info(`[Meteo] vigilance alerts parsed: ${alerts.length} for dept=${dept}`);
+        logger.info(`[MeteoFrance] vigilance alerts parsed: ${alerts.length} for dept=${dept}`);
       } catch (e) {
         const detail = e instanceof Error ? e.message : `${e}`;
-        logger.warn(`[Meteo] vigilance fetch failed for dept=${dept}: ${detail}`);
+        logger.warn(`[MeteoFrance] vigilance fetch failed for dept=${dept}: ${detail}`);
       }
     }
 
     res.json({
-      source,
-      house: { name: house.name },
       forecast: forecastData,
       vigilance: { alerts, dept, text },
     });
   }
 
   return {
-    'get /api/v1/service/meteo/vigilance': {
+    'get /api/v1/service/meteofrance/vigilance': {
       authenticated: true,
       controller: asyncMiddleware(getVigilanceController),
     },
-    'get /api/v1/service/meteo/vigilance/map': {
+    'get /api/v1/service/meteofrance/vigilance/map': {
       authenticated: true,
       controller: asyncMiddleware(getVigilanceMapController),
     },
-    'get /api/v1/house/:house_selector/meteo/weather': {
+    'get /api/v1/house/:house_selector/meteofrance/weather': {
       authenticated: true,
       controller: asyncMiddleware(getHouseWeatherController),
     },
