+++ server/services/meteofrance/lib/forecast.formatter.js
