+++ server/services/meteofrance/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "gladys-meteo",
+  "name": "gladys-meteofrance",
   "main": "index.js",
   "os": [
     "darwin",
